handwriting-ml-dj/pyproject.toml
[tool.poetry]
name = "handwriting-ml-dj"
version = "0.1.0"
description = ""
authors = ["rkhaq <rqk.email2@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Django = "3.2"
Pillow = "^9.1.0"
tensorflow = "^2.8.0"
opencv-python = "^4.5.5"
boto3 = "^1.22.9"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
